+++ lib/rubocop/cop/performance_cops.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative 'mixin/regexp_metacharacter'
+
 require_relative 'performance/bind_call'
 require_relative 'performance/caller'
 require_relative 'performance/case_when_splat'